+++ crates/openshell-server/tests/auth_endpoint_integration.rs
@@ -11,6 +11,8 @@
 //! a full `ServerState`.  The test handler mirrors the production logic in
 //! `auth.rs` but uses a simple `SocketAddr` as state.
 
+mod common;
+
 use axum::{
     Router,
     extract::{Query, State},
@@ -25,7 +27,6 @@ use hyper_util::rt::TokioIo;
 use serde::Deserialize;
 use std::net::SocketAddr;
 use tokio::net::TcpListener;
-use tonic::{Response, Status};
 
 // ---------------------------------------------------------------------------
 // Test auth handler (mirrors auth.rs but no ServerState)
@@ -379,437 +380,7 @@ async fn auth_connect_falls_back_to_bind_address() {
     server.abort();
 }
 
-// ---------------------------------------------------------------------------
-// Minimal OpenShell for test 7 (plaintext gRPC+HTTP)
-// ---------------------------------------------------------------------------
-
-#[derive(Clone, Default)]
-struct TestOpenShell;
-
-#[tonic::async_trait]
-impl openshell_core::proto::open_shell_server::OpenShell for TestOpenShell {
-    async fn health(
-        &self,
-        _request: tonic::Request<openshell_core::proto::HealthRequest>,
-    ) -> Result<Response<openshell_core::proto::HealthResponse>, Status> {
-        Ok(Response::new(openshell_core::proto::HealthResponse {
-            status: openshell_core::proto::ServiceStatus::Healthy.into(),
-            version: "test".to_string(),
-        }))
-    }
-
-    async fn create_sandbox(
-        &self,
-        _: tonic::Request<openshell_core::proto::CreateSandboxRequest>,
-    ) -> Result<Response<openshell_core::proto::SandboxResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::SandboxResponse::default(),
-        ))
-    }
-
-    async fn get_sandbox(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetSandboxRequest>,
-    ) -> Result<Response<openshell_core::proto::SandboxResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::SandboxResponse::default(),
-        ))
-    }
-
-    async fn list_sandboxes(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListSandboxesRequest>,
-    ) -> Result<Response<openshell_core::proto::ListSandboxesResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::ListSandboxesResponse::default(),
-        ))
-    }
-
-    async fn list_sandbox_providers(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListSandboxProvidersRequest>,
-    ) -> Result<Response<openshell_core::proto::ListSandboxProvidersResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::ListSandboxProvidersResponse::default(),
-        ))
-    }
-
-    async fn attach_sandbox_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::AttachSandboxProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::AttachSandboxProviderResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::AttachSandboxProviderResponse::default(),
-        ))
-    }
-
-    async fn detach_sandbox_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::DetachSandboxProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::DetachSandboxProviderResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::DetachSandboxProviderResponse::default(),
-        ))
-    }
-
-    async fn delete_sandbox(
-        &self,
-        _: tonic::Request<openshell_core::proto::DeleteSandboxRequest>,
-    ) -> Result<Response<openshell_core::proto::DeleteSandboxResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::DeleteSandboxResponse { deleted: true },
-        ))
-    }
-
-    async fn get_sandbox_config(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetSandboxConfigRequest>,
-    ) -> Result<Response<openshell_core::proto::GetSandboxConfigResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::GetSandboxConfigResponse::default(),
-        ))
-    }
-
-    async fn get_gateway_config(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetGatewayConfigRequest>,
-    ) -> Result<Response<openshell_core::proto::GetGatewayConfigResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::GetGatewayConfigResponse::default(),
-        ))
-    }
-
-    async fn get_sandbox_provider_environment(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetSandboxProviderEnvironmentRequest>,
-    ) -> Result<Response<openshell_core::proto::GetSandboxProviderEnvironmentResponse>, Status>
-    {
-        Ok(Response::new(
-            openshell_core::proto::GetSandboxProviderEnvironmentResponse::default(),
-        ))
-    }
-
-    async fn create_ssh_session(
-        &self,
-        _: tonic::Request<openshell_core::proto::CreateSshSessionRequest>,
-    ) -> Result<Response<openshell_core::proto::CreateSshSessionResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::CreateSshSessionResponse::default(),
-        ))
-    }
-
-    async fn expose_service(
-        &self,
-        _: tonic::Request<openshell_core::proto::ExposeServiceRequest>,
-    ) -> Result<Response<openshell_core::proto::ServiceEndpointResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::ServiceEndpointResponse::default(),
-        ))
-    }
-
-    async fn get_service(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetServiceRequest>,
-    ) -> Result<Response<openshell_core::proto::ServiceEndpointResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn list_services(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListServicesRequest>,
-    ) -> Result<Response<openshell_core::proto::ListServicesResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn delete_service(
-        &self,
-        _: tonic::Request<openshell_core::proto::DeleteServiceRequest>,
-    ) -> Result<Response<openshell_core::proto::DeleteServiceResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn revoke_ssh_session(
-        &self,
-        _: tonic::Request<openshell_core::proto::RevokeSshSessionRequest>,
-    ) -> Result<Response<openshell_core::proto::RevokeSshSessionResponse>, Status> {
-        Ok(Response::new(
-            openshell_core::proto::RevokeSshSessionResponse::default(),
-        ))
-    }
-
-    async fn create_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::CreateProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::ProviderResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn get_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::ProviderResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn list_providers(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListProvidersRequest>,
-    ) -> Result<Response<openshell_core::proto::ListProvidersResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn list_provider_profiles(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListProviderProfilesRequest>,
-    ) -> Result<Response<openshell_core::proto::ListProviderProfilesResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn get_provider_profile(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetProviderProfileRequest>,
-    ) -> Result<Response<openshell_core::proto::ProviderProfileResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn import_provider_profiles(
-        &self,
-        _: tonic::Request<openshell_core::proto::ImportProviderProfilesRequest>,
-    ) -> Result<Response<openshell_core::proto::ImportProviderProfilesResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn lint_provider_profiles(
-        &self,
-        _: tonic::Request<openshell_core::proto::LintProviderProfilesRequest>,
-    ) -> Result<Response<openshell_core::proto::LintProviderProfilesResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn delete_provider_profile(
-        &self,
-        _: tonic::Request<openshell_core::proto::DeleteProviderProfileRequest>,
-    ) -> Result<Response<openshell_core::proto::DeleteProviderProfileResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn update_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::UpdateProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::ProviderResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-    async fn get_provider_refresh_status(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetProviderRefreshStatusRequest>,
-    ) -> Result<Response<openshell_core::proto::GetProviderRefreshStatusResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn configure_provider_refresh(
-        &self,
-        _: tonic::Request<openshell_core::proto::ConfigureProviderRefreshRequest>,
-    ) -> Result<Response<openshell_core::proto::ConfigureProviderRefreshResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn rotate_provider_credential(
-        &self,
-        _: tonic::Request<openshell_core::proto::RotateProviderCredentialRequest>,
-    ) -> Result<Response<openshell_core::proto::RotateProviderCredentialResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn delete_provider_refresh(
-        &self,
-        _: tonic::Request<openshell_core::proto::DeleteProviderRefreshRequest>,
-    ) -> Result<Response<openshell_core::proto::DeleteProviderRefreshResponse>, Status> {
-        Err(Status::unimplemented("unused"))
-    }
-
-    async fn delete_provider(
-        &self,
-        _: tonic::Request<openshell_core::proto::DeleteProviderRequest>,
-    ) -> Result<Response<openshell_core::proto::DeleteProviderResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    type WatchSandboxStream = tokio_stream::wrappers::ReceiverStream<
-        Result<openshell_core::proto::SandboxStreamEvent, Status>,
-    >;
-    type ExecSandboxStream = tokio_stream::wrappers::ReceiverStream<
-        Result<openshell_core::proto::ExecSandboxEvent, Status>,
-    >;
-    type ConnectSupervisorStream = tokio_stream::wrappers::ReceiverStream<
-        Result<openshell_core::proto::GatewayMessage, Status>,
-    >;
-
-    async fn watch_sandbox(
-        &self,
-        _: tonic::Request<openshell_core::proto::WatchSandboxRequest>,
-    ) -> Result<Response<Self::WatchSandboxStream>, Status> {
-        let (_tx, rx) = tokio::sync::mpsc::channel(1);
-        Ok(Response::new(tokio_stream::wrappers::ReceiverStream::new(
-            rx,
-        )))
-    }
-
-    async fn exec_sandbox(
-        &self,
-        _: tonic::Request<openshell_core::proto::ExecSandboxRequest>,
-    ) -> Result<Response<Self::ExecSandboxStream>, Status> {
-        let (_tx, rx) = tokio::sync::mpsc::channel(1);
-        Ok(Response::new(tokio_stream::wrappers::ReceiverStream::new(
-            rx,
-        )))
-    }
-
-    type ExecSandboxInteractiveStream = tokio_stream::wrappers::ReceiverStream<
-        Result<openshell_core::proto::ExecSandboxEvent, Status>,
-    >;
-    async fn exec_sandbox_interactive(
-        &self,
-        _: tonic::Request<tonic::Streaming<openshell_core::proto::ExecSandboxInput>>,
-    ) -> Result<Response<Self::ExecSandboxInteractiveStream>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn update_config(
-        &self,
-        _: tonic::Request<openshell_core::proto::UpdateConfigRequest>,
-    ) -> Result<Response<openshell_core::proto::UpdateConfigResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn get_sandbox_policy_status(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetSandboxPolicyStatusRequest>,
-    ) -> Result<Response<openshell_core::proto::GetSandboxPolicyStatusResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn list_sandbox_policies(
-        &self,
-        _: tonic::Request<openshell_core::proto::ListSandboxPoliciesRequest>,
-    ) -> Result<Response<openshell_core::proto::ListSandboxPoliciesResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn report_policy_status(
-        &self,
-        _: tonic::Request<openshell_core::proto::ReportPolicyStatusRequest>,
-    ) -> Result<Response<openshell_core::proto::ReportPolicyStatusResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn get_sandbox_logs(
-        &self,
-        _: tonic::Request<openshell_core::proto::GetSandboxLogsRequest>,
-    ) -> Result<Response<openshell_core::proto::GetSandboxLogsResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn push_sandbox_logs(
-        &self,
-        _: tonic::Request<tonic::Streaming<openshell_core::proto::PushSandboxLogsRequest>>,
-    ) -> Result<Response<openshell_core::proto::PushSandboxLogsResponse>, Status> {
-        Err(Status::unimplemented("test"))
-    }
-
-    async fn submit_policy_analysis(
-        &self,
-        _request: tonic::Request<openshell_core::proto::SubmitPolicyAnalysisRequest>,
-    ) -> Result<Response<openshell_core::proto::SubmitPolicyAnalysisResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn get_draft_policy(
-        &self,
-        _request: tonic::Request<openshell_core::proto::GetDraftPolicyRequest>,
-    ) -> Result<Response<openshell_core::proto::GetDraftPolicyResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn approve_draft_chunk(
-        &self,
-        _request: tonic::Request<openshell_core::proto::ApproveDraftChunkRequest>,
-    ) -> Result<Response<openshell_core::proto::ApproveDraftChunkResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn reject_draft_chunk(
-        &self,
-        _request: tonic::Request<openshell_core::proto::RejectDraftChunkRequest>,
-    ) -> Result<Response<openshell_core::proto::RejectDraftChunkResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn approve_all_draft_chunks(
-        &self,
-        _request: tonic::Request<openshell_core::proto::ApproveAllDraftChunksRequest>,
-    ) -> Result<Response<openshell_core::proto::ApproveAllDraftChunksResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn edit_draft_chunk(
-        &self,
-        _request: tonic::Request<openshell_core::proto::EditDraftChunkRequest>,
-    ) -> Result<Response<openshell_core::proto::EditDraftChunkResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn undo_draft_chunk(
-        &self,
-        _request: tonic::Request<openshell_core::proto::UndoDraftChunkRequest>,
-    ) -> Result<Response<openshell_core::proto::UndoDraftChunkResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn clear_draft_chunks(
-        &self,
-        _request: tonic::Request<openshell_core::proto::ClearDraftChunksRequest>,
-    ) -> Result<Response<openshell_core::proto::ClearDraftChunksResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn get_draft_history(
-        &self,
-        _request: tonic::Request<openshell_core::proto::GetDraftHistoryRequest>,
-    ) -> Result<Response<openshell_core::proto::GetDraftHistoryResponse>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    async fn connect_supervisor(
-        &self,
-        _request: tonic::Request<tonic::Streaming<openshell_core::proto::SupervisorMessage>>,
-    ) -> Result<Response<Self::ConnectSupervisorStream>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    type RelayStreamStream =
-        tokio_stream::wrappers::ReceiverStream<Result<openshell_core::proto::RelayFrame, Status>>;
-
-    async fn relay_stream(
-        &self,
-        _request: tonic::Request<tonic::Streaming<openshell_core::proto::RelayFrame>>,
-    ) -> Result<Response<Self::RelayStreamStream>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-
-    type ForwardTcpStream = std::pin::Pin<
-        Box<
-            dyn tokio_stream::Stream<Item = Result<openshell_core::proto::TcpForwardFrame, Status>>
-                + Send,
-        >,
-    >;
-
-    async fn forward_tcp(
-        &self,
-        _request: tonic::Request<tonic::Streaming<openshell_core::proto::TcpForwardFrame>>,
-    ) -> Result<Response<Self::ForwardTcpStream>, Status> {
-        Err(Status::unimplemented("not implemented in test"))
-    }
-}
+use common::TestOpenShell;
 
 /// Test 7: Plaintext server (no TLS) accepts both gRPC and HTTP.
 ///
